Final Commit
Calculator is a simple project build on node.js using HTML5 and
angular.js in the client side. This application is used to perform some
basic mathematical functionalities like addition, subtraction,
multiplication and division. User is prompted with 2 input fields where
number is to be entered on these numbers, desired calculation is
performed. The functionality user wants to use can be chosen from one of
the 4 buttons on the page.